+++ libupnp/upnp/src/inc/httpreadwrite.h
@@ -494,8 +494,8 @@ int http_OpenHttpGetEx(IN const char *url_str,
 		     IN OUT char **contentType,
 		     OUT int *contentLength,
 		     OUT int *httpStatus,
-			 IN int lowRange,
-			 IN int highRange,
+			 IN unsigned long long lowRange,
+			 IN unsigned long long highRange,
 		     IN int timeout);
 
 /************************************************************************
